+++ src/parsers/common_args.js
@@ -8,6 +8,7 @@ import parseStyle from './style';
 export default function parseCommonArgs(target, mode) {
   const classes = parseClasses(target);
   classes.push(...classToArray(target[mode + 'Class'] || ''));
+  classes.sort();
   const transform = parseTransform(target.transform);
   const styles = parseStyle(target[mode + 'Style']);
   return {classes, transform, styles};