+++ src/utils/args_to_json.js
@@ -5,11 +5,7 @@ export default function argsToJSON(args) {
     if(this === args && val && (key === 'icon' || key === 'mask')) {
       return { prefix: val.prefix, iconName: val.iconName };
     }
-    if(this === args && key === 'classes' && Array.isArray(val)) {
-      return val.slice().sort();
-    }
     return val;
   }
-
   return sortedJSONStringify(args, replacer);
 }